seperated interruptions from simulator event loop
